% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/add_platt.R
\name{add_platt}
\alias{add_platt}
\alias{get_plattcoef}
\alias{get_plattcoef,tskrr-method}
\title{Add/get platt scaling coefs to/from a tskrr object}
\usage{
add_platt(x)

get_plattcoef(x)

\S4method{add_platt}{tskrr}(x)

\S4method{get_plattcoef}{tskrr}(x)
}
\arguments{
\item{x}{an object inheriting from \code{\link[xnet:tskrr-class]{tskrr}}}
}
\value{
the same object, but with the platt scaling coefficients filled in.
}
\description{
This function adds Platt scaling coefficients to any object inheriting from tskrr.
}
