% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/adjacencyData.R
\name{adjacencyData}
\alias{adjacencyData}
\alias{adjacencyData,matrix-method}
\title{Constructors for adjacencyData objects}
\usage{
adjacencyData(x, ...)

\S4method{adjacencyData}{matrix}(x)
}
\description{
These functions can be used to manually construct objects from the \code{\link[=dataClasses]{gramData classes}} for use in \code{\link{tskrr}} objects.
}
