% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_matrix.R
\name{as.matrix}
\alias{as.matrix}
\alias{as.matrix,}
\alias{adjacencyData-method}
\title{Convert data objects to matrix}
\usage{
\S4method{as.matrix}{adjacencyData}(x, ...)
}
\arguments{
\item{x}{an \R object.}

\item{...}{additional arguments to be passed to or from methods.}
}
\value{
a matrix
}
\description{
These functions let you convert \code{\link[xnet:dataClasses]{data classes}} to a matrix.
}
