% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/as_tuned.R
\docType{methods}
\name{as_tuned}
\alias{as_tuned}
\alias{as_tuned,tskrrHomogenous-method}
\alias{as_tuned,tskrrHeterogenous-method}
\title{convert tskrr model to a tuned model}
\usage{
as_tuned(x, ...)

\S4method{as_tuned}{tskrrHomogenous}(x, ...)

\S4method{as_tuned}{tskrrHeterogenous}(x, ...)
}
\arguments{
\item{x}{a model of class \code{\link[xnet:tskrr-class]{tskrr}}}

\item{...}{values for the extra slots defined by
the class \code{\link[xnet:tskrrTune-class]{tskrrTune}}}
}
\value{
a \code{\link[xnet:tskrrTune-class]{tskrrTune}} object of
the proper class (homogenous or heterogenous)
}
\description{
This function allows to convert a \code{\link[xnet:tskrr-class]{tskrr}}
model to a \code{\link[xnet:tskrrTune-class]{tskrrTune}} model of
the same type. The arguments take the values of the slots needed
for the conversion. More information on these slots can be found
at the help page of \code{\link[xnet:tskrrTune-class]{tskrrTune}}. This function is not exported.
}
\section{\bold{Warning}}{

This function does NOT tune a model. It is used to
construct the appropriate object by \code{\link{tune}}.
}

\seealso{
\itemize{
\item \code{\link{tune}} for actually tuning a model.
\item \code{\link[xnet:tskrrTune-class]{tskrrTune}} for
names and possible values of the slots passed through
\dots
}
}
