% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/convert.R
\name{convert}
\alias{convert}
\alias{convert,matrix,adjacencyDataMatrix-method}
\alias{convert,igraph,adjacencyDataIgraph-method}
\alias{convert,matrix,gramDataMatrix-method}
\alias{convert,data.frame,gramDataFrame-method}
\title{Carry out conversions}
\usage{
convert(x, using, ...)

\S4method{convert}{matrix,adjacencyDataMatrix}(x, using, args = NULL)

\S4method{convert}{igraph,adjacencyDataIgraph}(x, using, args = NULL)

\S4method{convert}{matrix,gramDataMatrix}(x, using, args = NULL)

\S4method{convert}{data.frame,gramDataFrame}(x, using, args = NULL)
}
\arguments{
\item{x}{the object to be converted}

\item{using}{the \code{\link{gramData}} or \code{\link{adjacencyData}} object to be used.}

\item{...}{extra parameters passed to other methods.}

\item{args}{a named list with extra arguments to be passed to the conversion function.}
}
\value{
the result of the conversion.
}
\description{
Convert an object using the conversion stored in one of
the \code{\link[=dataClasses]{data classes}}.
}
