% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_dataClasses.R
\docType{class}
\name{dataClasses}
\alias{dataClasses}
\alias{adjacencyData-class}
\alias{adjacencyDataIgraph-class}
\alias{adjacencyDataMatrix-class}
\alias{gramData-class}
\alias{gramDataMatrix-class}
\alias{gramDataFrame-class}
\title{Classes containing the original data}
\description{
The classes defined here serve as objects to store and
manipulate the original data used for creating a
\code{\link{tskrr}} model. They contain the original
data structure and the functions to transform the
original data to an adjacency matrix or a gram matrix.
}
\details{
The classes \code{adjacencyData} and \code{gramData} are virtual classes, and shouldn't be used directly.

The classes inheriting from \code{adjacencyData} can store the original data used to create an adjacency matrix. They also contain the matrix itself, and the conversion function used to construct it. The following classes inherit from \code{adjacencyData}:

\describe{
  \item{adjacencyDataMatrix}{the slot \code{adjacency} contains an adjacency matrix. The \code{conversion} functions is the \code{\link[base]{identity}} function by default.}
  \item{adjacencyDataIgraph}{the slot \code{orig} contains an \code{\link[igraph]{igraph}} object. The \code{conversion} function is either \code{\link[igraph]{as_adjacency_matrix}} or \code{\link[igraph]{as_incidence_matrix}}, depending on whether it's a homogeneous or heterogeneous network.}
}

The following classes inherit from \code{gramData}:

\describe{
  \item{gramDataMatrix}{BLAHBLAHBLAH}
}
}
\section{Slots}{

\describe{
\item{\code{orig}}{The original data. For adjacency objects, this
can be an \code{\link{igraph}} object. For gram objects,
this can be a matrix or a data frame.}

\item{\code{adjacency}}{The adjacency matrix used in the model (only for classes inheriting from \code{adjacencyData})}

\item{\code{conversion}}{A function used for the conversion of}

\item{\code{hasorig}}{a logical value, indicating whether a \code{DataMatrix} objects contain an original matrix.
In case the original matrix is already an adjacency or gram matrix,
this is set to \code{FALSE} and the slot \code{adjacency} or \code{gram} is  used as the original instead.}

\item{\code{eigen}}{The eigen-decomposition of the calculated
gram matrix (only for classes inheriting from \code{gramData})}

\item{\code{gram}}{The calculated gram matrix if \code{hasgram} is \code{TRUE} (only for classes inheriting from \code{gramData}).
the original data object to the adjacency or gram matrix.}

\item{\code{hasgram}}{a logical value indicating whether the gram matrix
is contained in the object or not.  (only for classes inheriting from \code{gramData}).}
}}

\seealso{
\code{\link{gramData}} for constructing \code{gramData} objects.
}
