% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{alpha}
\title{Deprecated functions}
\usage{
alpha(x)
}
\arguments{
\item{x}{an object.}
}
\value{
See the actual functions
}
\description{
The functions mentioned here will be removed in
future versions of the package.
}
\seealso{
for \code{alpha}: \code{\link{get_alpha}}
}
