% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim.R
\name{dim_tskrr}
\alias{dim_tskrr}
\alias{dim.tskrr}
\alias{dim,tskrr-method}
\alias{dim,gramData-method}
\title{Get the dimensions of an object}
\usage{
\S4method{dim}{tskrr}(x)

\S4method{dim}{gramData}(x)
}
\arguments{
\item{x}{a \code{\link[=tskrr-class]{tskrr}} or \code{\link[=dataClasses]{gramData}} object.}
}
\value{
a vector with two values indicating the number of rows
and the number of columns, or in case of a \code{gramData} object, a single number indicating the number of eigenvalues. This is equal to the number of rows/columns in the original gram matrix.
}
\description{
These functions allow you to extract the dimensions of a tskrr object or an object of one of the data classes.
}
\examples{
data(drugtarget)
mod <- tskrr(drugTargetInteraction, targetSim, drugSim)
dim(mod)
nrow(mod)
ncol(mod)

}
