% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_feature_importance.R
\docType{class}
\name{feature_importance-class}
\alias{feature_importance-class}
\title{Class feature_importance}
\description{
This class represents the feature importance outcomes. See also the function \code{\link{feature_importance}}.
}
\section{Slots}{

\describe{
\item{\code{orig_loss}}{a numeric value with the original loss of the model.}

\item{\code{n}}{the number of bootstrap samples}

\item{\code{loss_function}}{the function used to calculate the losses.}

\item{\code{exclusion}}{a character value indicating the exclusion setting used for the test.}

\item{\code{replaceby0}}{a logical value that indicates whether the excluseion was done by replacing with zero. See also \code{\link{loo}}}

\item{\code{features}}{the labels for the features that were checked.}

\item{\code{simloss}}{a matrix with the simulated losses for the bootstrapped samples}
}}

\seealso{
* the function \code{\link{feature_importance}} for the actual test
 * the function \code{\link{loo}} for the leave one out procedures
 @md

 @include all_generics.R

 @importFrom stats fivenum

 @rdname feature_importance-class
 @name feature_importance-class
 @exportClass feature_importance
}
