% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.R
\name{fitted}
\alias{fitted}
\title{extract the predictions}
\usage{
fitted(object, ...)
}
\arguments{
\item{object}{an object for which the extraction of model fitted values
is meaningful.}

\item{...}{other arguments.}
}
\description{
This functions extracts the fitted predictions from a
\code{\link[xnet:tskrr-class]{tskrr}} object. The \code{xnet}
package provides a generic for the \code{stats} function
\code{\link[stats]{fitted}} and a method for the
\code{\link[xnet:tskrr-class]{tskrr}} objects.
}
