% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetParametersTSKRR.R
\name{GetParametersTSKRR}
\alias{GetParametersTSKRR}
\title{Calculate the weights}
\usage{
GetParametersTSKRR(Y, U, Sigma, V, S, lambda.u, lambda.v)
}
\arguments{
\item{Y}{a matrix with the vertices (i.e. observed links)}

\item{U}{a matrix with eigenvectors for the rows}

\item{Sigma}{a vector with eigenvalues for the rows}

\item{V}{a matrix with eigenvectors for the columns}

\item{S}{a vector with eigenvalues for the columns}

\item{lambda.u}{a numeric value giving the lambda parameter for the rows}

\item{lambda.v}{a numeric value giginv the lambda parameter for the columns}
}
\value{
a matrix with the weights
}
\description{
This function returns the weight matrix for two-step kernel ridge
regression given eigenvalue decomposition of the Gram matrices.
}
\section{Warning}{

This is initial test code. It is currently NOT guaranteed to give
meaningful results.
}

