% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/getters_data.R,
%   R/getters_tskrr.R
\name{get_eigen}
\alias{get_eigen}
\alias{has_platt}
\alias{get_kernelmatrix,gramData-method}
\alias{getters-tskrr}
\alias{response}
\alias{response,tskrr-method}
\alias{lambda,tskrrHomogeneous-method}
\alias{lambda,tskrrHeterogeneous-method}
\alias{lambda}
\alias{is_tskrr}
\alias{is_homogeneous}
\alias{is_heterogeneous}
\alias{symmetry}
\alias{get_eigen,tskrr-method}
\alias{get_kernelmatrix,tskrr-method}
\alias{get_kernelmatrix}
\alias{has_hat}
\alias{has_platt,tskrr-method}
\title{Getters for tskrr objects}
\usage{
get_eigen(x, ...)

has_platt(x, ...)

\S4method{get_kernelmatrix}{gramData}(x)

\S4method{response}{tskrr}(x, ...)

\S4method{lambda}{tskrrHomogeneous}(x)

\S4method{lambda}{tskrrHeterogeneous}(x)

is_tskrr(x)

is_homogeneous(x)

is_heterogeneous(x)

symmetry(x)

\S4method{get_eigen}{tskrr}(x, which = c("row", "column"))

\S4method{get_kernelmatrix}{tskrr}(x, which = c("row", "column"))

has_hat(x)

\S4method{has_platt}{tskrr}(x)
}
\arguments{
\item{x}{a \code{\link[xnet:tskrr-class]{tskrr}} object or an
object inheriting from \code{tskrr}.}

\item{...}{arguments passed to other methods.}

\item{which}{a character value indicating whether the eigen decomposition
for the row kernel matrix or the column kernel matrix should be returned.}
}
\value{
For \code{response}: the original label matrix

For \code{lambda}: a named numeric vector with one resp both lambda
values used in the model. The names are "k" and "g" respectively.

For \code{is_tskrr} a logical value indicating whether the
object is a \code{tskrr} object

For \code{is_homogeneous} a logical value indicating whether the
tskrr model is a homogeneous one.

For \code{is_heterogeneous} a logical value indicating whether the
tskrr model is a heterogeneous one.

For \code{symmetry} a character value indicating the symmetry
for a \code{\link[xnet:tskrrHomogeneous-class]{homogeneous model}}. If
the model is not homogeneous, \code{NA} is returned.

For \code{get_eigen} the eigen decomposition of the requested
kernel matrix.

For \code{get_kernelmatrix} the original kernel matrix
for the rows or columns.

For \code{has_hat} a logical value indicating whether
the tskrr model contains the kernel hat matrices.

for \code{has_platt} a logical value indicating whether or not the object has coefficients for Platt scores
}
\description{
The functions described here are convenience functions to get
information out of a \code{\link[xnet:tskrr-class]{tskrr}} object.
}
\examples{
data(drugtarget)

mod <- tskrr(drugTargetInteraction, targetSim, drugSim)
is_homogeneous(mod)

EigR <- get_eigen(mod)
EigC <- get_eigen(mod, which = 'column')
lambda(mod)

}
