% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_tskrrImpute.R
\name{is_imputed}
\alias{is_imputed}
\alias{which_imputed}
\alias{imputed}
\title{Getters for tskrrImpute objects}
\usage{
is_imputed(x)

which_imputed(x)

imputed(x)
}
\arguments{
\item{x}{a \code{\link[xnet:tskrrImpute-class]{tskrrImpute}} object or
an object inheriting from \code{tskrrImpute}.}

\item{...}{arguments passed to other methods/}
}
\value{
For \code{is_imputed}: a logical value indicating whether
the model has imputed values

For \code{which_imputed}: a integer vector with the positions
for which the values are imputed.

for \code{imputed}: a matrix of the same dimensions as the
adjacency matrix. It contains the value \code{FALSE} at positions that
were not imputed, and \code{TRUE} at positions that were.
}
\description{
The functions described here are convenience functions to get
information out of a \code{\link[xnet:tskrrImpute-class]{tskrrImpute}}
object.
}
\examples{

data(drugtarget)

mod <- tskrr(drugTargetInteraction, targetSim, drugSim)

# NEED MORE EXAMPLES!

}
