% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_tskrrTune.R
\name{is_tuned}
\alias{is_tuned}
\title{Getters for tskrrTune objects}
\usage{
is_tuned(x)
}
\arguments{
\item{x}{a \code{\link[xnet:tskrrTune-class]{tskrrTune}} object or an
object inheriting from \code{tskrrTune}.}

\item{...}{arguments passed to other methods.}
}
\description{
The functions described here are convenience functions to get
information out of a \code{\link[xnet:tskrrTune-class]{tskrrTune}}
object.
}
