% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_linearFilter.R
\docType{methods}
\name{mean.linearFilter}
\alias{mean.linearFilter}
\alias{mean,linearFilter-method}
\title{Getters for linearFilter objects}
\usage{
\method{mean}{linearFilter}(x, ...)

\S4method{mean}{linearFilter}(x, ...)
}
\arguments{
\item{x}{a \code{linearFilter} object}
}
\value{
for \code{\bold{mean}}: the mean of the original matrix
}
\description{
These functions allow you to extract slots from objects of the
class \code{\link{linearFilter}}
}
