% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/gramData.R
\name{gramData}
\alias{gramData}
\alias{gramData,data.frame,function-method}
\alias{gramData,data.frame,ANY-method}
\title{Constructors for gramData objects}
\usage{
gramData(x, conversion, ...)

\S4method{gramData}{data.frame,`function`}(x, conversion, keep.gram = FALSE)

\S4method{gramData}{data.frame,ANY}(
  x,
  conversion,
  kernel = rbfdot,
  kernel.args = list(),
  keep.gram = FALSE
)
}
\arguments{
\item{x}{an object to be converted to an \code{adjacencyData} or \code{gramData} object}

\item{conversion}{a conversion function or a formula specifying which variables should be used to be converted. See Details}

\item{...}{arguments passed to other methods.}

\item{keep.gram}{a logical value indicating whether the gram matrix should be kept or not.}

\item{kernel}{a kernel generating function from the package \code{\link[kernlab:dots]{kernlab}}. See details.}

\item{kernel.args}{a list with further arguments to be passed to \code{kernel}.}
}
\description{
These functions can be used to manually construct objects from the \code{\link[=dataClasses]{gramData classes}} for use in \code{\link{tskrr}} objects.
}
\details{
When constructing a \code{gramData} object from a \code{data.frame}, you can either specify a conversion function yourself, or use a kernel from the package \code{kernlab}. If you select one of these kernels, the data frame is first converted to a model matrix according to the specified formula (see \code{\link[stats]{model.matrix}}). On this matrix, the chosen kernel is then applied. By default, a Gaussian RBF kernel with \code{sigma = 1} is used on a model matrix formed by all variables and without an intercept.
}
\seealso{
\code{\link{dataClasses}} for an overview of the different classes generated by these functions.
}
