% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\docType{methods}
\name{importance,tskrr-method}
\alias{importance,tskrr-method}
\alias{importance,tskrrTune-method}
\title{Calculate the relative importance of the edges}
\usage{
\S4method{importance}{tskrr}(x, permutation = c("both", "row", "column"),
  n = 100, exclusion = c("interaction", "row", "column", "both"))

\S4method{importance}{tskrrTune}(x, permutation = c("both", "row",
  "column"), n = 100)
}
\arguments{
\item{x}{either a \code{\link{tskrr-class}} or a
\code{\link{tskrrTune-class}} object}
}
\value{
a matrix with the differences in loss
}
\description{
This function does a permutation-based evaluation of the impact of
different edges on the final result.
}
