% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\docType{methods}
\name{labels.tskrr}
\alias{labels.tskrr}
\alias{labels,tskrr-method}
\alias{dimnames,tskrr-method}
\alias{dimnames.tskrr}
\alias{rownames,tskrr-method}
\alias{colnames,tskrr-method}
\title{Extract labels from a tskrr object}
\usage{
\method{labels}{tskrr}(object, prefix = if (is_homogenous(object)) "row" else
  c("row", "col"), ...)

\S4method{labels}{tskrr}(object, prefix = if (is_homogenous(object)) "row"
  else c("row", "col"), ...)

\S4method{dimnames}{tskrr}(x)

\S4method{rownames}{tskrr}(x, do.NULL = TRUE, prefix = "row")

\S4method{colnames}{tskrr}(x, do.NULL = TRUE, prefix = "col")
}
\arguments{
\item{object}{a \code{\link{tskrr}} object}

\item{prefix}{a prefix used for construction of the labels in case
none are available. For \code{label}, a character vector of length 1 or 2.
In case two values are given, the first is used for the rows and the second
for the columns. Otherwise the only value is used for both. In the case of
\code{rownames} and \code{colnames}, a single value.
See also \code{\link[=row+colnames]{row+colnames}}}

\item{...}{arguments passed to/from other methods.}

\item{x}{a \code{\link{tskrr}} object}
}
\value{
for \code{labels} and \code{dimnames}: a list with two elements \code{k} and
\code{g}
}
\description{
These functions allow you to extract the labels from a
\code{\link{tskrr}} object. The function \code{labels} and the
function \code{dimnames} are aliases and do the exact same
thing.
}
\details{
The functions \code{colnames} and \code{rownames} work like you would
expect, with one difference: the rownames or colnames for a \code{tskrr}
object can never be \code{NULL}.
}
