% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_filter.R
\name{linear_filter}
\alias{linear_filter}
\title{Fit a linear filter over an adjacency matrix}
\usage{
linear_filter(y, alpha = 0.25)
}
\arguments{
\item{y}{an adjacency matrix}

\item{alpha}{a vector with 4 alpha values, or a single alpha value
which then is used for all 4 alphas.}
}
\value{
an object of class \code{\link[class-linearFilter]{linearFilter}}
}
\description{
This function fits a linear filter over an adjacency matrix.
}
\details{
NEEDS DIFFERENT OUTPUT AND ALPHA CONTROL
}
\examples{
data(drugtarget)
linear_filter(drugTargetInteraction, alpha = 0.25)
linear_filter(drugTargetInteraction, alpha = c(0.1,0.1,0.4,0.4))

}
