% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_linearFilter.R
\docType{class}
\name{linearFilter-class}
\alias{linearFilter-class}
\alias{linearFilter}
\title{Class linearFilter}
\description{
The class represents the outcome of a linear filter, and is normally
generated by the function \code{\link{linear_filter}}
}
\section{Slots}{

\describe{
\item{\code{y}}{the original adjacency matrix with responses.}

\item{\code{alpha}}{a numeric vector with the 4 alpha values of the model.}

\item{\code{pred}}{a matrix with the predictions}

\item{\code{mean}}{a numeric vector containing the global mean of \code{y}}

\item{\code{colmeans}}{a numeric vector containing the column means of \code{y}}

\item{\code{rowmeans}}{a numeric vector containing the row means of \code{y}.}
}}

\seealso{
\code{\link{linear_filter}} for creating a linear filter model,
and \code{\link[=getters_linearFilter]{getter fuctions for linearFilter}}.
}
