% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/looInternal.R
\name{loo_internal}
\alias{loo_internal}
\alias{loo.i}
\alias{loo.i0}
\alias{loo.r}
\alias{loo.c}
\alias{loo.b}
\alias{loo.e.sym}
\alias{loo.e.skew}
\alias{loo.e0.sym}
\alias{loo.e0.skew}
\alias{loo.v}
\title{Leave-one-out crossvalidation for two-step kernel ridge regression}
\usage{
loo.i(Y, Hk, Hg, pred)

loo.i0(Y, Hk, Hg, pred)

loo.r(Y, Hk, Hg)

loo.c(Y, Hk, Hg)

loo.b(Y, Hk, Hg)

loo.e.sym(Y, Hk, pred)

loo.e.skew(Y, Hk, pred)

loo.e0.sym(Y, Hk, pred)

loo.e0.skew(Y, Hk, pred)

loo.v(Y, Hk)
}
\arguments{
\item{Y}{the matrix with responses}

\item{Hk}{the hat matrix for the first kernel}

\item{Hg}{the hat matrix for the second kernel}

\item{pred}{the predictions}
}
\value{
a matrix with the leave-one-out predictions
}
\description{
These functions implement different cross-validation scenarios for
two-step kernel ridge regression. It uses the shortcuts for
leave-one-out crossvalidation.
}
\details{
These functions are primarily for internal use and hence not exported
}
