% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/permtest.R
\docType{methods}
\name{permtest}
\alias{permtest}
\alias{permtest,tskrrHomogenous-method}
\alias{permtest,tskrrTune-method}
\title{Calculate the relative importance of the edges}
\usage{
permtest(x, ...)

\S4method{permtest}{tskrrHeterogenous}(x, permutation = c("both", "row",
  "column"), n = 100, exclusion = c("interaction", "row", "column",
  "both"), replaceby0 = FALSE, fun = loss_mse)

\S4method{permtest}{tskrrHomogenous}(x, permutation = c("both"),
  n = 100, exclusion = c("interaction", "both"), replaceby0 = FALSE,
  fun = loss_mse)

\S4method{permtest}{tskrrTune}(x, permutation = c("both", "row",
  "column"), n = 100)
}
\arguments{
\item{x}{either a \code{\link{tskrr-class}} or a
\code{\link{tskrrTune-class}} object}

\item{permutation}{a character string that defines whether the row,
column or both kernel matrices should be permuted. Ignored in case of
a homogenous network}

\item{n}{the number of permutations for every kernel matrix}

\item{exclusion}{the exclusion to be used in the \code{\link{loo}} function. See also \code{\link{get_loo_fun}}}

\item{replaceby0}{a logical value indicating whether \code{\link{loo}}
removes a value in the leave-one-out procedure or replaces it by zero.
See also \code{\link{get_loo_fun}}.}

\item{fun}{a function (or a character string with the name of a
function) that calculates the loss. See also \code{\link{tune}} and
\code{\link{loss_functions}}}
}
\value{
a matrix with the differences in loss
}
\description{
This function does a permutation-based evaluation of the impact of
different edges on the final result.
}
