% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platt_scores.R
\name{platt_scores}
\alias{platt_scores}
\alias{platt_scores,tskrr-method}
\title{Calculate Platt scores}
\usage{
platt_scores(x, ...)

\S4method{platt_scores}{tskrr}(x)
}
\arguments{
\item{x}{an object that contains or represents the outcome}
}
\value{
Depending on the input, a vector or matrix with the platt scores.
}
\description{
Platt scaling uses a logistic regression to convert the linear output of a two-step kernel ridge regression to approximate probabilities.
}
