% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\docType{methods}
\name{predict.tskrr}
\alias{predict.tskrr}
\alias{predict,tskrr-method}
\title{predict method for tskrr fits}
\usage{
predict.tskrr(object, K, G = NULL, ...)

\S4method{predict}{tskrr}(object, K, G = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{\link[xnet:tskrr-class]{tskrr}}.}

\item{K}{a new K matrix}

\item{G}{a new G matrix or \code{NULL}. If \code{NULL}, K is used
for both.}

\item{...}{arguments passed to or from other methods

This function is currently in testing and should be used
with caution.}
}
\description{
Obtains predictions from a \code{\link{tskrr}} model for new
data. To get the predictions on the training data, use the
function \code{\link[xnet:fitted]{fitted}}
}
