% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReconstrucFromEigen.R
\name{ReconstructFromEigen}
\alias{ReconstructFromEigen}
\title{Reconstruct from eigen}
\usage{
ReconstructFromEigen(vectors, values)
}
\arguments{
\item{vectors}{a matrix with eigenvectors}

\item{values}{a vector with eigenvalues}
}
\value{
a matrix
}
\description{
This function "reconstructs" from the eigen decomposition.
}
\note{
This function is not exported.
}
\section{Warning}{

This is initial test code. It is currently NOT guaranteed to give
meaningful results.
}

