% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrainTSKRR.R
\name{TrainTSKRR}
\alias{TrainTSKRR}
\title{Train a two-step kernel ridge regression}
\usage{
TrainTSKRR(Y, K, G, lambda.u = 1, lambda.v = 1)
}
\arguments{
\item{Y}{a matrix with the observations}

\item{K}{a matrix - Gram matrix for the rows}

\item{G}{a matrix - Gram matrix for the columns}

\item{lambda.u}{a numeric value for the lambda parameter for the rows}

\item{lambda.v}{a numeric value for the lambda parameter for the columns}
}
\value{
a list with the model components. Currently classless
}
\description{
This function trains the two step kernel ridge regression
}
\section{Warning}{

This is initial test code. It is currently NOT guaranteed to give
meaningful results.
}

