% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tskrr.fit.R
\name{tskrr.fit}
\alias{tskrr.fit}
\title{Carry out a two-step kernel ridge regression}
\usage{
tskrr.fit(y, k, g = NULL, lambda.k = NULL, lambda.g = NULL, ...)
}
\arguments{
\item{y}{a matrix representing the links between the nodes of both
networks.}

\item{k}{an object of class \code{\link{eigen}} containing the eigen
decomposition of the first kernel matrix.}

\item{g}{an optional object of class \code{\link{eigen}} containing
the eigen decomposition of the second kernel matrix. If \code{NULL},
the network is considered to be homogenous.}

\item{lambda.k}{a numeric value for the lambda parameter tied
to the first kernel.}

\item{lambda.g}{a numeric value for the lambda parameter tied
to the second kernel. If \code{NULL}, the model is fit using the same
value for \code{lambda.k} and \code{lambda.g}}

\item{...}{arguments passed to other functions. Currently ignored.}
}
\value{
a list with three elements:
\itemize{
   \item k : the hat matrix for the rows
   \item g : the hat matrix for the columns (or \code{NULL})
   for homogenous networks.
   \item pred : the predictions
}
}
\description{
This function provides an interface for two-step kernel ridge regression.
To use this function, you need at least one kernel matrix and one
interaction matrix. It's the internal engine used by the function
\code{\link{tskrr}}.
}
\details{
This function is mostly available for internal use. In most cases it
makes much more sense to use \code{\link{tskrr}}, as that function
returns an object one can actually work with. The function
\code{tskrr.fit} could be useful when doing simulations or
fitting algorithms, as the information returned from this function
is enough to use the functions returned by \code{\link{get_loo_fun}}.
}
\examples{

data(drugtarget)

K <- eigen(targetSim)
G <- eigen(drugSim)

res <- tskrr.fit(drugTargetInteraction,K,G,
                 lambda.k = 0.01, lambda.g = 0.05)

}
