% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tskrr.R
\name{tskrr}
\alias{tskrr}
\title{Fitting a two step kernel ridge regression}
\usage{
tskrr(y, k, g, homogenous = NULL)
}
\arguments{
\item{y}{a matrix}

\item{k}{a matrix}

\item{g}{a matrix}

\item{homogenous}{a logical value indicating whether the fitting should
be done for a homogenous network. If \code{NULL} the function will
derive the type of network form the other inputs. Defaults to \code{NULL}.}
}
\value{
a \code{\link[xnet:tskrr-class]{tskrr}} object
}
\description{
\code{tskrr} is the main function for fitting a two step kernel
ridge regression. It can be used for both homogenous and heterogenous
networks.
}
\details{
THIS FUNCTION IS NOT WORKING YET.
}
