% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_tskrrHeterogeneous.R
\docType{class}
\name{tskrrHeterogeneous-class}
\alias{tskrrHeterogeneous-class}
\alias{tskrrHeterogeneous}
\title{Class tskrrHeterogeneous}
\description{
The class tskrrHeterogeneous is a subclass of the superclass
\code{\link[xnet:tskrr-class]{tskrr}} specifically for
heterogeneous networks.
}
\section{Slots}{

\describe{
\item{\code{y}}{an adjacency matrix.}

\item{\code{k}}{an object of class \code{\link[=dataClasses]{gramData}}}

\item{\code{lambda.k}}{the lambda value used for k}

\item{\code{pred}}{the matrix with the predictions}

\item{\code{g}}{an object of class \code{\link[=dataClasses]{gramData}}}

\item{\code{lambda.g}}{the lambda value used for g}

\item{\code{has.hat}}{a logical value indicating whether the kernel hat matrices
are stored in the object.}

\item{\code{Hk}}{the kernel hat matrix for the rows.}

\item{\code{Hg}}{the kernel hat matrix for the columns.}
}}

