% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_tskrrHeterogenous.R
\docType{class}
\name{tskrrHeterogenous-class}
\alias{tskrrHeterogenous-class}
\alias{tskrrHeterogenous}
\title{Class tskrrHeterogenous}
\description{
The class tskrrHeterogenous is a subclass of the superclass
\code{\link[xnet:tskrr-class]{tskrr}} class specifically for
heterogenous networks.
}
\section{Slots}{

\describe{
\item{\code{y}}{the matrix with responses}

\item{\code{k}}{the eigen decomposition of the kernel matrix for the rows}

\item{\code{lambda.k}}{the lambda value used for k}

\item{\code{pred}}{the matrix with the predictions}

\item{\code{g}}{the eigen decomposition of the kernel matrix for the columns}

\item{\code{lambda.g}}{the lambda value used for g}

\item{\code{has.hat}}{a logical value indicating whether the kernel hat matrices
are stored in the object.}

\item{\code{Hk}}{the kernel hat matrix for the rows.}

\item{\code{Hg}}{the kernel hat matrix for the columns.}

\item{\code{labels}}{a list with elements \code{k} and \code{g} (see
\code{\link{tskrr-class}}).
 If any element is \code{NA}, the labels used
are integers indicating the row resp column number.}
}}

