% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_tskrrHomogeneous.R
\docType{class}
\name{tskrrHomogeneous-class}
\alias{tskrrHomogeneous-class}
\alias{tskrrHomogeneous}
\title{Class tskrrHomogeneous}
\description{
The class tskrrHomogeneous is a subclass of the superclass
\code{\link[xnet:tskrr-class]{tskrr}} specifically for
homogeneous networks.
}
\section{Slots}{

\describe{
\item{\code{y}}{an adjacency matrix.}

\item{\code{k}}{an object of class \code{\link[=dataClasses]{gramData}}}

\item{\code{lambda.k}}{the lambda value used for k}

\item{\code{pred}}{the matrix with the predictions}

\item{\code{symmetry}}{a character value that can have the possible values
\code{"symmetric"}, \code{"skewed"} or \code{"not"}. It indicates
whether the \code{y} matrix is symmetric, skewed-symmetric or not
symmetric.}

\item{\code{has.hat}}{a logical value indicating whether the kernel hat matrices
are stored in the object.}

\item{\code{Hk}}{the kernel hat matrix for the rows.}
}}

