% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_tskrrHomogeneousImpute.R
\docType{class}
\name{tskrrHomogeneousImpute-class}
\alias{tskrrHomogeneousImpute-class}
\alias{tskrrHomogeneousImpute}
\title{Class tskrrHomogeneousImpute}
\description{
The class \code{tskrrHomogeneousImpute} is a subclass of the
class \code{\link[xnet:tskrrHomogeneous-class]{tskrrHomogeneous}} and
\code{\link[xnet:tskrrImpute-class]{tskrrImpute}}
specifically for homogeneous networks with imputed values. It is
the result of the function \code{\link{impute_tskrr}} on a
homogeneous network model.
}
\section{Slots}{

\describe{
\item{\code{y}}{the matrix with responses}

\item{\code{k}}{the eigen decomposition of the kernel matrix for the rows}

\item{\code{lambda.k}}{the lambda value used for k}

\item{\code{pred}}{the matrix with the predictions}

\item{\code{symmetry}}{a character value that can have the possible values
\code{"symmetric"}, \code{"skewed"} or \code{"not"}. It indicates
whether the \code{y} matrix is symmetric, skewed-symmetric or not
symmetric.}

\item{\code{has.hat}}{a logical value indicating whether the kernel hat matrices
are stored in the object.}

\item{\code{Hk}}{the kernel hat matrix for the rows.}

\item{\code{labels}}{a list with elements \code{k} and \code{g} (see
\code{\link{tskrr-class}}). For homogeneous networks, \code{g}
is always \code{NA}. If \code{k} is \code{NA}, the labels used
are integers indicating the row resp column number.}

\item{\code{imputeid}}{a vector with integer values indicating which of
the values in \code{y} are imputed}

\item{\code{niter}}{an integer value gving the number of iterations used}

\item{\code{tol}}{a numeric value with the tolerance used}
}}

