% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_tskrrHomogenous.R
\docType{class}
\name{tskrrHomogenous-class}
\alias{tskrrHomogenous-class}
\alias{tskrrHomogenous}
\title{Class tskrrHomogenous}
\description{
The class tskrrHomogenous is a subclass of the superclass
\code{\link[xnet:tskrr-class]{tskrr}} class specifically for
homogenous networks.
}
\section{Slots}{

\describe{
\item{\code{y}}{the matrix with responses}

\item{\code{k}}{the eigen decomposition of the kernel matrix for the rows}

\item{\code{lambda.k}}{the lambda value used for k}

\item{\code{pred}}{the matrix with the predictions}

\item{\code{symmetry}}{a character value that can have the possible values
\code{"symmetric"}, \code{"skewed"} or \code{"not"}. It indicates
whether the \code{y} matrix is symmetric, skewed-symmetric or not
symmetric.}

\item{\code{has.orig}}{a logical value indicating whether the original kernel
matrices are stored in the object.}

\item{\code{k.orig}}{the original kernel matrix for the rows.}
}}

