% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_tskrrImpute.R
\docType{class}
\name{tskrrImpute-class}
\alias{tskrrImpute-class}
\alias{tskrrImpute}
\title{Class tskrrImpute}
\description{
The class \code{tskrrImpute} is a virtual class that represents a
\code{\link[xnet:tskrr-class]{tskrr}} model with imputed values in
the adjacency matrix Y. Apart from the model, it contains extra
information on the imputed values.
}
\section{Slots}{

\describe{
\item{\code{imputeid}}{a vector with integer values indicating which of
the values in \code{y} are imputed}

\item{\code{imputemethod}}{a character value with either \code{"predictions"} or
\code{"loo"} indicating whether the predictions or the leave-one-out
crossvalidation values were used in the imputation.}

\item{\code{loo_settings}}{in case \code{imputemethod = "loo"}, a list with
two elements named \code{exclusion} and \code{replaceby0}. These
indicate the arguments used in \code{\link{get_loo_fun}}.}

\item{\code{niter}}{an integer value gving the number of iterations used}

\item{\code{tol}}{a numeric value with the tolerance used}
}}

