% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_tskrrTuneHeterogeneous.R
\docType{class}
\name{tskrrTuneHeterogeneous-class}
\alias{tskrrTuneHeterogeneous-class}
\alias{tskrrTuneHeterogeneous}
\title{Class tskrrTuneHeterogeneous}
\description{
The class tskrrTuneHeterogeneous represents a tuned Heterogeneous
\code{\link[xnet:tskrr-class]{tskrr}} model. It inherits from
the classes \code{\link[xnet:tskrrHeterogeneous-class]{tskrrHeterogeneous}}
and \code{\link[xnet:tskrrTune-class]{tskrrTune}}.
}
