% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_tskrrTuneHeterogenous.R
\docType{class}
\name{tskrrTuneHeterogenous-class}
\alias{tskrrTuneHeterogenous-class}
\alias{tskrrTuneHeterogenous}
\title{Class tskrrTuneHeterogenous}
\description{
The class tskrrTuneHeterogenous represents a tuned Heterogenous
\code{\link[xnet:tskrr-class]{tskrr}} model. It inherits from
the classes \code{\link[xnet:tskrrHeterogenous-class]{tskrrHeterogenous}}
and \code{\link[xnet:tskrrTune-class]{tskrrTune}}.
}
