% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_tskrrTuneHomogeneous.R
\docType{class}
\name{tskrrTuneHomogeneous-class}
\alias{tskrrTuneHomogeneous-class}
\alias{tskrrTuneHomogeneous}
\title{Class tskrrTuneHomogeneous}
\description{
The class tskrrTuneHomogeneous represents a tuned homogeneous
\code{\link[xnet:tskrr-class]{tskrr}} model. It inherits from
the classes \code{\link[xnet:tskrrHomogeneous-class]{tskrrHomogeneous}}
and \code{\link[xnet:tskrrTune-class]{tskrrTune}}.
}
