% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_tskrrTuneHomogenous.R
\docType{class}
\name{tskrrTuneHomogenous-class}
\alias{tskrrTuneHomogenous-class}
\alias{tskrrTuneHomogenous}
\title{Class tskrrTuneHomogenous}
\description{
The class tskrrTuneHomogenous represents a tuned homogenous
\code{\link[xnet:tskrr-class]{tskrr}} model. It inherits from
the classes \code{\link[xnet:tskrrHomogenous-class]{tskrrHomogenous}}
and \code{\link[xnet:tskrrTune-class]{tskrrTune}}.
}
