% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune.R
\docType{methods}
\name{tune}
\alias{tune}
\alias{tune,tskrrHomogenous-method}
\alias{tune,tskrrHeterogenous-method}
\title{tune the lambda parameters for a tskrr}
\usage{
\S4method{tune}{tskrrHomogenous}(x, lim = c(1e-04, 10000), ngrid = 10,
  lambda = NULL)

\S4method{tune}{tskrrHeterogenous}(x, lim = c(1e-04, 10000), ngrid = 10,
  lambda = NULL)
}
\arguments{
\item{x}{a \code{\link{tskrr}} object representing a two step
kernel ridge regression model.}

\item{lim}{a vector with 2 values that give the boundaries for the domain
in which lambda is searched, or possibly a list with 2 elements. See details}

\item{ngrid}{a single numeric value giving the number of points
in a single dimension of the grid, or possibly a list with 2 elements.
See details.}

\item{lambda}{a vector with the lambdas that need checking for
homogenous networks, or possibly a list with two elements for
heterogenous networks. See Details. Defaults to
\code{NULL}, which means that the function constructs the search grid
from the other arguments.}
}
\description{
This function lets you tune the lambda parameter(s) of a two-step
kernel ridge regression for optimal performance. The objective
function used for tuning is the squared norm based on leave-one-out
crossvalidation.
}
\details{
This function currently only performs a naive grid search for all
(combinations of) lambda values. If no specific lambda values are
provided, then the function uses \code{\link{create_grid}} to
create a grid that's evenly spaced on a logarithmic scale.

In the case of a heterogenous network, you can specify different values
for the two parameters that need tuning. In order to do so, you need to
provide a list with the settings for every parameter to the arguments
\code{lim}, \code{ngrid} and/or \code{lambda}. If you
try this for a homogenous network, the function will return an error.
}
