% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/update.R
\docType{methods}
\name{update}
\alias{update}
\alias{update,tskrrHomogenous-method}
\alias{update,tskrrHeterogenous-method}
\title{Update a tskrr object with a new lambda}
\usage{
update(object, ...)

\S4method{update}{tskrrHomogenous}(object, lambda)

\S4method{update}{tskrrHeterogenous}(object, lambda)
}
\arguments{
\item{object}{a \code{\link[xnet:tskrr-class]{tskrr}} object}

\item{...}{arguments passed to methods}

\item{lambda}{a numeric vector with one or two values for the
hyperparameter lambda. If two values are given, the first one is
used for the k matrix and the second for the g matrix.}
}
\value{
an updated \code{\link[xnet:tskrr-class]{tskrr}} object
fitted with the new lambdas.
}
\description{
This function allows you to refit a \code{\link{tskrr}} with a
new lambda. It can be used to do manual tuning/crossvalidation.
}
