% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_dimensions.R
\name{valid_dimensions}
\alias{valid_dimensions}
\alias{is_square}
\title{Functions to check dimension}
\usage{
valid_dimensions(y, k, g = NULL)

is_square(x)
}
\arguments{
\item{y}{a matrix}

\item{k}{a gramData object or matrix}

\item{g}{an optional second gramData object or matrix, or \code{NULL} otherwise.}

\item{x}{any matrix}
}
\value{
a logical value indicating whether the dimensions of the
objects are compatible for a two step kernel ridge regression.
}
\description{
These functions allow you to check whether the dimensions of the
adjacency matrix and the \code{\link{gramData}} objects are compatible.
\code{valid_dimensions} checks whether y has as many rows as k and whether y has as many columns as g.
\code{is_square} checks whether both dimensions are the same.
}
\note{
The function \code{is_square} is not exported
}
\seealso{
\code{\link{gramData}} for conversions.
}
