% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xnet-package.R
\docType{package}
\name{xnet-package}
\alias{xnet}
\alias{xnet-package}
\title{Two-step kernel ridge regression for network analysis}
\description{
This package implements two-step kernel ridge regression, a technique
that can be used for all kinds of cross-network analysis. Examples are
protein-protein interaction, foodwebs, ...
}
\seealso{
Send your bug reports to:

\url{https://github.com/CenterForStatistics-UGent/xnet/issues}

More background in the paper by Stock et al, 2018:

\url{http://doi.org/10.1093/bib/bby095}
}
\author{
Joris Meys and Michiel Stock
}
