% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/as_tuned.R
\name{as_tuned}
\alias{as_tuned}
\alias{as_tskrr}
\alias{as_tuned,tskrrHomogeneous-method}
\alias{as_tuned,tskrrHeterogeneous-method}
\alias{as_tskrr,tskrrTune-method}
\alias{as_tskrr,tskrrImpute-method}
\alias{as_tskrr,tskrr-method}
\title{convert tskrr models}
\usage{
as_tuned(x, ...)

as_tskrr(x, ...)

\S4method{as_tuned}{tskrrHomogeneous}(x, ...)

\S4method{as_tuned}{tskrrHeterogeneous}(x, ...)

\S4method{as_tskrr}{tskrrTune}(x)

\S4method{as_tskrr}{tskrrImpute}(x)

\S4method{as_tskrr}{tskrr}(x)
}
\arguments{
\item{x}{a model of class \code{\link[xnet:tskrr-class]{tskrr}}}

\item{...}{values for the extra slots defined by
the class \code{\link[xnet:tskrrTune-class]{tskrrTune}}}
}
\value{
For \code{as_tuned}:
a \code{\link[xnet:tskrrTune-class]{tskrrTune}} object of
the proper class (homogeneous or heterogeneous)

For \code{as_tskrr}: an object of class
\code{\link[xnet:tskrrHomogeneous-class]{tskrrHomogeneous}} or
\code{\link[xnet:tskrrHeterogeneous-class]{tskrrHeterogeneous}} depending
on whether the original object was homogeneous or heterogeneous.
}
\description{
These functions allow converting models that inherit from the
\code{\link[xnet:tskrr-class]{tskrr}} and
\code{\link[xnet:tskrrTune-class]{tskrrTune}} class into each other,
keeping track of whether the model is homogeneous or heterogeneous.
The dots argument allows specifying values for possible extra slots
when converting from \code{tskrr} to \code{tskrrTune}.
More information on these slots can be found
on the help page of \code{\link[xnet:tskrrTune-class]{tskrrTune}}.
\strong{These functions are not exported.}
}
\section{\bold{Warning}}{

This functions do NOT tune a model. they are used internally to
make the connection between both types in the methods.
}

\seealso{
\itemize{
\item \code{\link{tune}} for actually tuning a model.
\item \code{\link[xnet:tskrrTune-class]{tskrrTune}} for
names and possible values of the slots passed through
\dots
}
}
