% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/as_tuned.R
\docType{methods}
\name{as_tuned}
\alias{as_tuned}
\alias{as_tskrr}
\alias{as_tuned,tskrrHomogenous-method}
\alias{as_tuned,tskrrHeterogenous-method}
\alias{as_tskrr,tskrrTune-method}
\alias{as_tskrr,tskrr-method}
\title{convert tskrr models}
\usage{
as_tuned(x, ...)

as_tskrr(x, ...)

\S4method{as_tuned}{tskrrHomogenous}(x, ...)

\S4method{as_tuned}{tskrrHeterogenous}(x, ...)

\S4method{as_tskrr}{tskrrTune}(x)

\S4method{as_tskrr}{tskrr}(x)
}
\arguments{
\item{x}{a model of class \code{\link[xnet:tskrr-class]{tskrr}}}

\item{...}{values for the extra slots defined by
the class \code{\link[xnet:tskrrTune-class]{tskrrTune}}}
}
\value{
For \code{as_tuned}:
a \code{\link[xnet:tskrrTune-class]{tskrrTune}} object of
the proper class (homogenous or heterogenous)

For \code{as_tskrr}: an object of class
\code{\link[xnet:tskrrHomogenous-class]{tskrrHomogenous}} or
\code{\link[xnet:tskrrHeterogenous-class]{tskrrHeterogenous}} depending
on whether the original object was homogenous or heterogenous.
}
\description{
These functions allow to convert models that inherit from the
\code{\link[xnet:tskrr-class]{tskrr}} and
\code{\link[xnet:tskrrTune-class]{tskrrTune}} class into eachother,
keeping track of whether the model is homogenous or heterogenous.
The dots argument allows to specify values for possible extra slots
when converting from \code{tskrr} to \code{tskrrTune}.
More information on these slots can be found
at the help page of \code{\link[xnet:tskrrTune-class]{tskrrTune}}.
These functions are not exported.
}
\section{\bold{Warning}}{

This functions do NOT tune a model. they are used internally to
make the connection between both types in the methods.
}

\seealso{
\itemize{
\item \code{\link{tune}} for actually tuning a model.
\item \code{\link[xnet:tskrrTune-class]{tskrrTune}} for
names and possible values of the slots passed through
\dots
}
}
