% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/convert.R
\name{convert}
\alias{convert}
\alias{convert,matrix,gramDataMatrix-method}
\alias{convert,data.frame,gramDataFrame-method}
\title{Carry out conversions}
\usage{
convert(x, using, ...)

\S4method{convert}{matrix,gramDataMatrix}(x, using, args = NULL)

\S4method{convert}{data.frame,gramDataFrame}(x, using, args = NULL)
}
\arguments{
\item{x}{the object to be converted}

\item{using}{the \code{\link{gramData}} object to be used.}

\item{...}{extra parameters passed to other methods.}

\item{args}{a named list with extra arguments to be passed to the conversion function.}
}
\value{
an object inheriting from \code{matrix} with the gram matrix where the rows present the data in \code{x} and the columns the data in the original object.
}
\description{
Convert an object using the conversion stored in one of
the \code{\link[=dataClasses]{data classes}}. With this function you can
calculate the gram matrix after applying the same kernel as used by the
\code{\link{gramData}} object passed to the \code{using} argument.
This function is useful for predictions, as the returned objects are used
by the function \code{\link[=predict.tskrr]{predict}}
to calculate the predictions.
}
