% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_dataClasses.R
\docType{class}
\name{dataClasses}
\alias{dataClasses}
\alias{gramData-class}
\alias{gramDataMatrix-class}
\alias{gramDataFrame-class}
\title{gramData classes}
\description{
The classes defined here serve as objects to store and
manipulate the original data used for creating a
\code{\link{tskrr}} model. They can contain the original
data structure and the functions to transform the
original data to an a gram matrix.
}
\details{
The class \code{gramData} contains the eigendecomposition of the gram matrix. If requested, the gram matrix itself is stored in the object as well, and the slot \code{hasgram} is set to TRUE.

The classes inheriting from \code{gramData} can store the original data used to create a gram matrix, as well as the conversion function used for this. This is especially useful when a kernel function is applied to either a matrix or a data frame. The following classes inherit from \code{gramData}:

\describe{
  \item{gramDataMatrix}{This contains an original matrix and a conversion function to convert it to a gram matrix. Standard it would be some kind of kernel function.}
  \item{gramDataFrame}{This contains a data frame and a conversion function to convert it to a gram matrix. The function \code{\link{gramData}} can do this conversion based on a kernel function and a formula that indicates which variables should be used to calculate the gram matrix. }
}
}
\section{Slots}{

\describe{
\item{\code{orig}}{The original data, if provided.}

\item{\code{conversion}}{A function used for the conversion of the original object to either an adjacency matrix or a gram matrix.}

\item{\code{conv.args}}{a list with arguments to be used when calling the function in \code{conversion}.}

\item{\code{hasorig}}{a logical value, indicating whether objects contain an original matrix.}

\item{\code{symmetric}}{a logical value, indicating whether the adjacency matrix is a symmetric one (homogeneous networks) or not (heterogeneous networks).}

\item{\code{eigen}}{The eigen-decomposition of the calculated
gram matrix.}

\item{\code{gram}}{The calculated gram matrix if \code{hasgram} is \code{TRUE}.}

\item{\code{hasgram}}{a logical value indicating whether the gram matrix
is contained in the object or not.}

\item{\code{labels}}{a character vector with the labels for the gram matrix.}
}}

\seealso{
\code{\link{gramData}} for constructing \code{gramData} objects.
}
