% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen2hat.R
\name{eigen2hat}
\alias{eigen2hat}
\title{Calculate the hat matrix from an eigen decomposition}
\usage{
eigen2hat(eigen, vec, lambda)
}
\arguments{
\item{eigen}{a matrix with the eigenvectors.}

\item{vec}{an numeric vector with the eigen values.}

\item{lambda}{a single numeric value for the hyperparameter lambda}
}
\value{
a numeric matrix representing the hat matrix
}
\description{
This function calculates the hat matrix for a two-step kernel ridge
regression based on the eigen decomposition of the kernel matrix.
}
