% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_feature_importance.R, R/all_generics.R,
%   R/feature_importance.R
\name{print.feature_importance}
\alias{print.feature_importance}
\alias{feature_importance}
\alias{feature_importance,tskrr-method}
\title{Importance of features}
\usage{
\method{print}{feature_importance}(x, digits = max(3L, getOption("digits") - 3), ...)

feature_importance(x, ...)

\S4method{feature_importance}{tskrr}(
  x,
  n = 100,
  dimension = c("row", "column"),
  exclusion = "interaction",
  replaceby0 = FALSE,
  fun = loss_auc,
  features,
  verbose = interactive()
)
}
\arguments{
\item{x}{a \code{\link{tskrr}} oject.}

\item{digits}{the number of digits shown in the output}

\item{...}{arguments passed to or from other methods.}

\item{n}{a single integer indicating the number of bootstrap samples used to estimate the feature importance.}

\item{dimension}{a character value indicating for which kernel matrix the features should be evaluated (rows or columns). Ignored in the case of a homogeneous network.}

\item{exclusion}{a character value with possible values "interaction",
"row", "column", "both" for heterogeneous models, and "edges", "vertices",
"interaction" or "both" for homogeneous models.
Defaults to "interaction". See details.}

\item{replaceby0}{a logical value indicating whether the interaction
should be simply removed (\code{FALSE}) or replaced by 0 (\code{TRUE}).}

\item{fun}{a function to be used for calculating the loss. This
can also be a character value giving the name of one of the loss
functions provided in the package}

\item{features}{a vector with indices (numeric, character or logical) that selects the feature columns that need to be evaluated. If empty, the function will loop over all features.}

\item{verbose}{a logical value that turns on a message indicating which feature is processed. By default, it's turned on in interactive sessions.}
}
\description{
With this function you can calculate the importance of the features used to create the \code{\link{gramData}} object. This doesn't work if you provided the kernel matrices directly.
}
\details{
The function will select each feature in the original data frame in turn, and resample it \code{n} times. Each time the loss function is recalculated, and the difference with the original loss value is stored. These values can then later be summarized or plotted.

When a \code{\link{tskrr}} model was tuned earlier with the function \code{\link{tune}}, the values for \code{exclusion} and \code{replaceby0} are taken from the tuned object. Otherwise these arguments are passed to the function \code{\link{loss}} to calculate the loss value.
}
