% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.R
\docType{methods}
\name{fitted.tskrr}
\alias{fitted.tskrr}
\alias{fitted,tskrr-method}
\title{extract the predictions}
\usage{
\method{fitted}{tskrr}(object, ...)

\S4method{fitted}{tskrr}(object, ...)
}
\arguments{
\item{object}{an object for which the extraction of model fitted values
is meaningful.}

\item{...}{arguments passed to or from other methods.}
}
\value{
a numeric matrix with the predictions
}
\description{
This functions extracts the fitted predictions from a
\code{\link[xnet:tskrr-class]{tskrr}} object. The \code{xnet}
package provides a generic for the \code{stats} function
\code{\link[=fitted]{fitted}} and a method for the
\code{\link[xnet:tskrr-class]{tskrr}} objects.
}
\examples{

data(drugtarget)

mod <- tskrr(drugTargetInteraction, targetSim, drugSim)
pred <- fitted(mod)

}
