% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_loo_fun.R
\name{get_loo_fun}
\alias{get_loo_fun}
\title{Retrieve a loo function}
\usage{
get_loo_fun(exclusion, homogenous, symmetry, replaceby0)
}
\arguments{
\item{exclusion}{a character value with possible values "interaction",
"row", "column" or "both". Defaults to "interaction". See details.}

\item{homogenous}{a logical value indicating whether
this should be for a homogenous model or not.}

\item{symmetry}{a character value with possible values
\code{"symmetric"}, \code{"skewed"} or \code{"not"}. If
\code{homogenous = FALSE} this argument is ignored. See also
\code{\link{symmetry}} and the
\code{\link[xnet:tskrrHomogenous-class]{tskrrHomogenous class}}}

\item{replaceby0}{a logical value indicating whether the interaction
should be simply removed (\code{FALSE}) or replaced by 0 (\code{TRUE}).
See Details.}
}
\value{
a function taking the arguments y, Hr, and possibly pred
for calculating the leave one out crossvalidation. When
\code{homogenous = FALSE}, the returned function also
has an argument Hc.
}
\description{
This function returns the correct function needed to perform
one of the leave one out crossvalidations. It primarily meant
for internal use but can be handy when doing simulations.
}
\details{
This function can be used to select the correct loo function in
a simulation or tuning algorithm.
}
\seealso{
\code{\link{loo}} for carrying out a leave on out crossvalidation,
and \code{\link{loo_internal}} for more information on the internal
functions one retrieves with this one.
}
