% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_loo_fun.R
\name{get_loo_fun}
\alias{get_loo_fun}
\title{Find correct loo function}
\usage{
get_loo_fun(exclusion, homogenous, symmetry, replaceby0 = FALSE)
}
\arguments{
\item{exclusion}{a character value with possible values "interaction",
"row", "column" or "both". Defaults to "interaction". See details.}

\item{homogenous}{a logical value indicating whether or
not the network is homogenous}

\item{symmetry}{a character value with the possible values}

\item{replaceby0}{a logical value indicating whether the interaction
should be simply removed (\code{FALSE}) or replaced by 0 (\code{TRUE}).
This only makes sense when \code{exclusion = "interaction"} and the
response matrix contains only 0 and 1 values. In all other cases, setting
this value to \code{TRUE} will result in an error.}
}
\value{
the function needed to calculate the loo predictions
}
\description{
This function is a wrapper that looks for the correct internal
function to carry out a leave one out crossvalidation. The
function is not exported.
}
\details{
The parameter \code{exclusion} defines what exactly is left out.
The value "interaction" means that a single interaction is removed.
In the case of a homogenous model this can be interpreted as the
removal of the interaction between two edges. The values "row" and
"column" mean that all interactions for a row edge resp. a column
edge are removed. The value "both" removes all interactions for
a row and a column edge.

In the case of a homogenous model, "row" and "column" don't make sense
and will be replaced by "both". This can be interpreted
as removing vertices, i.e. all interactions between one edge and
all other edges. For more information, see Stock et al PAPER TO BE ADDED.
}
