% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/getters_data.R
\name{has_orig}
\alias{has_orig}
\alias{get_orig}
\alias{getters-data}
\alias{has_orig,adjacencyData-method}
\alias{has_orig,gramData-method}
\alias{get_orig,adjacencyData-method}
\alias{get_orig,gramData-method}
\title{Getters for the classes adjacencyData and gramData}
\usage{
has_orig(x, ...)

get_orig(x, ...)

\S4method{has_orig}{adjacencyData}(x)

\S4method{has_orig}{gramData}(x)

\S4method{get_orig}{adjacencyData}(x)

\S4method{get_orig}{gramData}(x)
}
\arguments{
\item{x}{an object of any of those classes or classes that inherit from those classes.}

\item{...}{arguments passed to other methods}
}
\description{
These functions allow you to get slots from the classes \code{\link[xnet:dataClasses]{adjacencyData}} and \code{\link[xnet:dataclasses]{gramData}}.
}
