% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Getters_tskrr.R
\docType{methods}
\name{response,tskrr-method}
\alias{response,tskrr-method}
\alias{response}
\alias{lambda,tskrr-method}
\alias{lambda,tskrrHeterogenous-method}
\alias{lambda}
\alias{is_homogenous}
\alias{symmetry}
\title{Getters for tskrr objects}
\usage{
\S4method{response}{tskrr}(x, ...)

\S4method{lambda}{tskrr}(x)

\S4method{lambda}{tskrrHeterogenous}(x)

is_homogenous(x)

symmetry(x)
}
\arguments{
\item{x}{a \code{\link[xnet:tskrr-class]{tskrr}} object or an
object inheriting from \code{tskrr}.}

\item{...}{arguments passed to other methods.}
}
\value{
For \code{response}: the original response matrix

For \code{lambda}: a named numeric vector with one resp both lambda
values used in the model. The names are "k" and "g" respectively.

For \code{is_homogenous} a logical value indicating whether the
tskrr model is a homogenous one.

For \code{symmetry} a character value indicating the symmetry
for a \code{\link[xnet:tskrrHomogenous-class]{homogenous model}}. If
the model is not homogenous, \code{NA} is returned.
}
\description{
The functions described here are convenience functions to get
information out of a \code{\link[xnet:tskrr-class]{tskrr}} object.
}
