% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Getters_tskrr.R
\docType{methods}
\name{response,tskrr-method}
\alias{response,tskrr-method}
\alias{response}
\alias{lambda,tskrr-method}
\alias{lambda,tskrrHeterogenous-method}
\alias{lambda}
\title{Getters for tskrr objects}
\usage{
\S4method{response}{tskrr}(x, ...)

\S4method{lambda}{tskrr}(x)

\S4method{lambda}{tskrrHeterogenous}(x)
}
\arguments{
\item{x}{a \code{\link[xnet:tskrr-class]{tskrr}} object or an
object inheriting from \code{tskrr}.}

\item{...}{arguments passed to other methods.}
}
\value{
For \code{response}: the original response matrix

For \code{lambda}: a named numeric vector with one resp both lambda
values used in the model. The names are "k" and "g" respectively.
}
\description{
The functions described here are convenience functions to get
information out of a \code{\link[xnet:tskrr-class]{tskrr}} object.
}
