% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hat.R
\docType{methods}
\name{hat,tskrrHeterogenous-method}
\alias{hat,tskrrHeterogenous-method}
\alias{hat,tskrrHomogenous-method}
\title{Return the hat matrix of a tskrr model}
\usage{
\S4method{hat}{tskrrHeterogenous}(x, which = c("row", "column"))

\S4method{hat}{tskrrHomogenous}(x, ...)
}
\arguments{
\item{x}{a tskrr model}

\item{which}{a character value with possible values "row" or
"column" to indicate which should be returned. For homogenous
models, this parameter is ignored.}

\item{...}{arguments passed to other methods.}
}
\value{
the requested hat matrix of the model.
}
\description{
This function returns the hat matrix or hat matrices of
a tskrr model. \code{xnet} creates an S4 generic for \code{hat}
and links the default method to the \code{\link[=influence.measures]{hat}} function
of \code{stats}
}
