% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/hat.R
\name{hat}
\alias{hat}
\alias{hat,tskrrHeterogeneous-method}
\alias{hat,tskrrHomogeneous-method}
\title{Return the hat matrix of a tskrr model}
\usage{
hat(x, ...)

\S4method{hat}{tskrrHeterogeneous}(x, which = c("row", "column"))

\S4method{hat}{tskrrHomogeneous}(x, ...)
}
\arguments{
\item{x}{a tskrr model}

\item{...}{arguments passed to other methods.}

\item{which}{a character value with possible values "row" or
"column" to indicate which should be returned. For homogeneous
models, this parameter is ignored.}
}
\value{
the requested hat matrix of the model.
}
\description{
This function returns the hat matrix or hat matrices of
a tskrr model. \code{xnet} creates an S4 generic for \code{hat}
and links the default method to the \code{\link[=influence.measures]{hat}} function
of \code{stats}
}
