% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/looInternal.R
\name{loo_internal}
\alias{loo_internal}
\alias{loo.i}
\alias{loo.i0}
\alias{loo.r}
\alias{loo.c}
\alias{loo.b}
\alias{loo.e.sym}
\alias{loo.e.skew}
\alias{loo.e0.sym}
\alias{loo.e0.skew}
\alias{loo.v}
\title{Leave-one-out crossvalidation for two-step kernel ridge regression}
\usage{
loo.i(Y, Hr, Hc, pred)

loo.i0(Y, Hr, Hc, pred)

loo.r(Y, Hr, Hc, ...)

loo.c(Y, Hr, Hc, ...)

loo.b(Y, Hr, Hc, ...)

loo.e.sym(Y, Hr, pred)

loo.e.skew(Y, Hr, pred)

loo.e0.sym(Y, Hr, pred)

loo.e0.skew(Y, Hr, pred)

loo.v(Y, Hr, ...)
}
\arguments{
\item{Y}{the matrix with responses}

\item{Hr}{the hat matrix for the first kernel (rows of Y)}

\item{Hc}{the hat matrix for the second kernel (columns of Y)}

\item{pred}{the predictions}

\item{...}{added to allow for specifying pred even when not needed.}
}
\value{
a matrix with the leave-one-out predictions
}
\description{
These functions implement different cross-validation scenarios for
two-step kernel ridge regression. It uses the shortcuts for
leave-one-out crossvalidation.
}
\details{
These functions are primarily for internal use and hence not exported.
Be careful when using them, as they do not perform any sanity check
on the input. It's up to the user to make sure the input makes sense.
}
\seealso{
\code{\link{loo}} for the user-level function.
}
