% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permtest.R
\docType{methods}
\name{permtest,tskrr-method}
\alias{permtest,tskrr-method}
\title{Calculate the relative importance of the edges}
\usage{
\S4method{permtest}{tskrr}(x, permutation = c("both", "row", "column"),
  n = 100, exclusion = c("interaction", "row", "column", "both"))
}
\arguments{
\item{x}{either a \code{\link{tskrr-class}} or a
\code{\link{tskrrTune-class}} object}
}
\value{
a matrix with the differences in loss
}
\description{
This function does a permutation-based evaluation of the impact of
different edges on the final result.
}
