% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_grid.R
\name{plot_grid}
\alias{plot_grid}
\title{Plot the grid of a tuned tskrr model}
\usage{
plot_grid(x, ...)
}
\arguments{
\item{x}{an object that inherits from
\code{\link[xnet:tskrrTune-class]{tskrrTune}}}

\item{...}{arguments passed to image}
}
\value{
\code{NULL} invisibly
}
\description{
With this function you can visualize the grid search for optimal
lambdas from a \code{\link[xnet:tskrrTune-class]{tskrrTune}} object.
In the case of a heterogenous network, this function will plot a
contour plot on a grid, based on the functions \code{\link{image}}
and \code{\link{contour}}
}
