% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_tskrr.R
\docType{class}
\name{tskrr-class}
\alias{tskrr-class}
\title{Class tskrr}
\description{
The class tskrr represents a two step kernel ridge regression fitting
object, and is normally generated by the function \code{\link{tskrr}}.
This is a superclass so it should not be instantiated directly.
}
\section{Slots}{

\describe{
\item{\code{y}}{an adjacency matrix}

\item{\code{k}}{an object of class \code{\link[=dataClasses]{gramData}}}

\item{\code{lambda.k}}{the lambda value used for k}

\item{\code{pred}}{the matrix with the predictions}

\item{\code{has.hat}}{a logical value indicating whether the kernel hat matrices
are stored in the object.}

\item{\code{Hk}}{the kernel hat matrix for the rows.}

\item{\code{platt}}{a numeric vector with 2 values containing the coefficients for Platt scaling (if applicable). If Platt scaling isn't applicable, it contains an empty vector.}
}}

\seealso{
the classes \code{\link{tskrrHomogeneous}} and
\code{\link{tskrrHeterogeneous}} for the actual classes.
}
