% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tskrr.fit.R
\name{tskrr.fit}
\alias{tskrr.fit}
\title{Carry out a two-step kernel ridge regression}
\usage{
tskrr.fit(y, k, g = NULL, lambda.k = NULL, lambda.g = NULL, ...)
}
\arguments{
\item{y}{a matrix representing the links between the nodes of both
networks.}

\item{k}{an object of class \code{\link{eigen}} containing the eigen
decomposition of the first kernel matrix.}

\item{g}{an optional object of class \code{\link{eigen}} containing
the eigen decomposition of the second kernel matrix. If \code{NULL},
the network is considered to be homogenous.}

\item{lambda.k}{a numeric value for the lambda parameter tied
to the first kernel.}

\item{lambda.g}{a numeric value for the lambda parameter tied
to the second kernel. If \code{NULL}, the model is fit using the same
value for \code{lambda.k} and \code{lambda.g}}

\item{...}{arguments passed to other functions. Currently ignored.}
}
\value{
TO BE DETERMINED
}
\description{
This function provides an interface for two-step kernel ridge regression.
To use this function, you need at least one kernel matrix and one
interaction matrix.
}
