% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tskrr.R
\name{tskrr}
\alias{tskrr}
\title{Fitting a two step kernel ridge regression}
\usage{
tskrr(y, k, g = NULL, lambda = 1e-04, homogenous = is.null(g),
  testdim = TRUE, keep = FALSE)
}
\arguments{
\item{y}{a response matrix}

\item{k}{a kernel matrix for the rows}

\item{g}{an optional kernel matrix for the columns}

\item{lambda}{a numeric vector with one or two values for the
hyperparameter lambda. If two values are given, the first one is
used for the k matrix and the second for the g matrix.}

\item{homogenous}{a logical value indicating whether the fitting should
be done for a homogenous network. Normally this is obvious from the
input (i.e. a lacking \code{g} matrix indicates the network is homogenous)}

\item{testdim}{a logical value indicating whether symmetry
and the dimensions of the kernel(s) should be tested.
Defaults to \code{TRUE}, but for large matrices
putting this to \code{FALSE} will speed up the function.}

\item{keep}{a logical value indicating whether the original kernel
matrices should be stored in the model object. Doing so makes the
model object quite larger, but can speed up predictions in
some cases. Defaults to \code{FALSE}}
}
\value{
a \code{\link[xnet:tskrr-class]{tskrr}} object
}
\description{
\code{tskrr} is the main function for fitting a two step kernel
ridge regression. It can be used for both homogenous and heterogenous
networks.
}
\details{
TO DO:

- add symmetry checks (for skewed or symmetric or none)
}
\examples{

# Heterogenous network

data(drugtarget)

mod <- tskrr(drugTargetInteraction, targetSim, drugSim)

Y <- response(mod)
pred <- fitted(mod)

# Homogenous network

data(proteinInteraction)

modh <- tskrr(proteinInteraction, Kmat_y2h_sc)

Yh <- response(modh)
pred <- fitted(modh)

}
\seealso{
\code{\link{response}}, \code{\link{fitted}},
\code{\link{get_eigen}}, \code{\link{eigen2hat}}
}
