% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_dimensions.R
\name{valid_dimensions}
\alias{valid_dimensions}
\alias{is_square}
\title{Functions to check matrices}
\usage{
valid_dimensions(y, k, g = NULL)

is_square(x)
}
\arguments{
\item{y}{a response matrix}

\item{k}{a kernel matrix}

\item{g}{an optional second kernel matrix or \code{NULL} otherwise.}

\item{x}{any matrix}
}
\value{
a logical value indicating whether the dimensions of the
matrices are compatible for a two step kernel ridge regression.
}
\description{
This functions allow you to check whether the dimensions of the
response matrix and the kernel matrix (matrices) are compatible.
\code{valid_dimensions} checks whether both k and g are square matrices,
whether y has as many rows as k and whether y has as many columns as g.
\code{is_square} just checks whether both dimensions are the same.
}
\note{
The function \code{is_square} is not exported
}
