% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{weights,tskrrHeterogeneous-method}
\alias{weights,tskrrHeterogeneous-method}
\alias{weights}
\alias{weights,tskrrHomogeneous-method}
\title{Extract weights from a tskrr model}
\usage{
\S4method{weights}{tskrrHeterogeneous}(object)

\S4method{weights}{tskrrHomogeneous}(object)
}
\arguments{
\item{object}{a \code{\link{tskrr}} object for which the weights
have to be calculated.}

\item{...}{other arguments passed to methods.}
}
\value{
a matrix with the weights for the tskrr model.
}
\description{
This function calculates the weight matrix for
calculating the predictions of a tskrr model.
}
\details{
The weight matrix is calculated from the map matrices through the
function \code{\link{eigen2map}}.
}
\note{
The package \code{xnet} adds a S4 generic function
for \code{\link[stats]{weights}}.
}
