% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xnet-package.R
\docType{package}
\name{xnet-package}
\alias{xnet}
\alias{xnet-package}
\title{Two-step kernel ridge regression for network analysis}
\description{
This package implements two-step kernel ridge regression, a technique
that can be used for all kinds of cross-network analysis. Examples are
protein-protein interaction, foodwebs, ...
}
